Refactoring:
move code related to individual models to separate packages,
including "multi-model"
